Compiler-warning cleanup and clang-tidy pass
Mechanical warning cleanup across the tree plus clang-tidy findings (header
guards, override specifiers, unused-parameter marks). Upstream PR #736 covers
the C side; the C++ batch is the planned follow-up.

+++ src/actors/fake_item_box/render.inc.c
@@ -26,7 +26,7 @@ void render_actor_fake_item_box(Camera* camera, struct FakeItemBox* fakeItemBox)
     f32 someMultiplier;
 
     size_t actorIdx = CM_FindActorIndex((struct Actor*)fakeItemBox);
-    if (-1 == actorIdx) {
+    if (-1 == (s32) actorIdx) {
         printf("[render_actor_fake_item_box] Could not find fake item box for FI, skipping!\n");
         return;
     }
